fix: resolve CI type errors, clippy warnings, and ESLint unused-imports
Contracts:
- Remove unnecessary 'as u32' cast (clippy was failing with -D warnings)

Backend:
- Update stellar.ts for @stellar/stellar-sdk v12 API: SorobanRpc.Operations
  → Operation, isSimulationError → Api.isSimulationError, Horizon.Asset → Asset,
  fix TransactionBuilder.sign → Transaction.sign pattern
- Fix getContractData calls in payroll.ts/stream.ts: v12 takes (contractId, key)
  as ScVal, not LedgerKey
- Fix import path in schemas.test.ts (wrong relative path)
- Add missing scValToNative import in payroll.ts

Frontend:
- Remove unused imports: Building2, Clock, result/setResult, useQuery,
  isPending, Play, Pause, CheckCircle

+++ contracts/payroll-manager/src/lib.rs
@@ -111,7 +111,7 @@ impl PayrollManager {
         token: Address,
     ) {
         admin.require_auth();
-        if min_signers == 0 || min_signers > signers.len() as u32 {
+        if min_signers == 0 || min_signers > signers.len() {
             panic!("invalid signer threshold");
         }
         let company = Company { admin: admin.clone(), signers, min_signers, token, active: true };
@@ -131,7 +131,7 @@ impl PayrollManager {
         if !company.active {
             panic!("company is deactivated");
         }
-        if min_signers == 0 || min_signers > signers.len() as u32 {
+        if min_signers == 0 || min_signers > signers.len() {
             panic!("invalid signer threshold");
         }
         company.signers = signers;
